+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+- Add experimental feature to detect resource detectors in auto instrumentation
+  ([#3181](https://github.com/open-telemetry/opentelemetry-python/pull/3181))
 - Fix exporting of ExponentialBucketHistogramAggregation from opentelemetry.sdk.metrics.view
   ([#3240](https://github.com/open-telemetry/opentelemetry-python/pull/3240))
 - Fix headers types mismatch for OTLP Exporters